requirements.txt
kaggle==1.5.3
Pillow==5.4.1
opencv-python==4.0.0.21
numpy==1.16.2
tensorflow-gpu==1.13.1
pathos==0.2.3
matplotlib==3.0.3
bokeh==1.0.4
scipy==1.2.1
seaborn==0.9.0
scikit-learn==0.20.3
